Decompose MainScreen into smaller, focused components
- Extract ImageComponents.kt: CameraSection, CapturedImageCard, SampleImagesSection
- Extract StateSections.kt: LoadingSection, ErrorSection, ErrorActionButtons
- Extract ResultsComponents.kt: MainResultsSection, RecipeResultsSection
- Extract LayoutComponents.kt: MainScreenHeader, InstructionalTextSection, AnalyzeButtonSection
- Reduce MainScreen.kt from 963 lines to ~250 lines
- Add androidTest fake repositories: FakeAIRepository, FakeSecurityRepository
- Add androidTest DI modules: TestAppModule, TestRepositoryModule
- Fix integration test dependencies for proper androidTest environment

Benefits:
- Improved maintainability and readability
- Better separation of concerns
- Enhanced reusability of components
- Easier testing and debugging
- Better team collaboration potential
